Replace `DosageAdapter` with the generic `ItemAdapter`
